+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # CHANGELOG
 
 ## [Unreleased]
+- add-image-panel: feat: add panel to load images from URL (Alexander Goussas)
 - add-cmd-panel: feat: add panel to display output of system command (Alexander Goussas)